+++ deduce/__init__.py
@@ -1,10 +1,5 @@
 import importlib.metadata
 
-from deduce.deduce import (
-    Deduce,
-    annotate_text,
-    annotate_text_structured,
-    deidentify_annotations,
-)
+from deduce.deduce import Deduce
 
 __version__ = importlib.metadata.version(__package__ or __name__)